refactor: move dom utils to utils folder
